+++ Civi/DataProcessor/FieldOutputHandler/ContactLinkFieldOutputHandler.php
@@ -86,7 +86,10 @@ class ContactLinkFieldOutputHandler extends AbstractFieldOutputHandler implement
         2=>$configuration['datasource'])
       ));
     }
-    $this->contactIdField = $this->contactIdSource->getAvailableFields()->getFieldSpecificationByName($configuration['contact_id_field']);
+    $this->contactIdField = $this->contactIdSource->getAvailableFields()->getFieldSpecificationByAlias($configuration['contact_id_field']);
+    if (!$this->contactIdField) {
+      $this->contactIdField = $this->contactIdSource->getAvailableFields()->getFieldSpecificationByName($configuration['contact_id_field']);
+    }
     if (!$this->contactIdField) {
       throw new FieldNotFoundException(E::ts("Field %1 requires a field with the name '%2' in the data source '%3'. Did you change the data source type?", array(
         1 => $title,
@@ -103,7 +106,10 @@ class ContactLinkFieldOutputHandler extends AbstractFieldOutputHandler implement
         2=>$configuration['contact_name_datasource'])
       ));
     }
-    $this->contactNameField = $this->contactNameSource->getAvailableFields()->getFieldSpecificationByName($configuration['contact_name_field']);
+    $this->contactNameField = $this->contactNameSource->getAvailableFields()->getFieldSpecificationByAlias($configuration['contact_name_field']);
+    if (!$this->contactNameField) {
+      $this->contactNameField = $this->contactNameSource->getAvailableFields()->getFieldSpecificationByName($configuration['contact_name_field']);
+    }
     if (!$this->contactNameField) {
       throw new FieldNotFoundException(E::ts("Field %1 requires a field with the name '%2' in the data source '%3'. Did you change the data source type?", array(
         1 => $title,